moved messaging classes into appropriate package
