+++ byon-agent/src/main/java/io/github/cloudiator/iaas/byon/messaging/ByonNodeQuerySubscriber.java
@@ -16,31 +16,31 @@
  * under the License.
  */
 
-package io.github.cloudiator.iaas.byon;
+
+package io.github.cloudiator.iaas.byon.messaging;
 
 import com.google.inject.Inject;
-import io.github.cloudiator.iaas.byon.util.IdCreator;
+import io.github.cloudiator.iaas.byon.Constants;
+import io.github.cloudiator.messaging.ByonToByonMessageConverter;
 import io.github.cloudiator.persistance.ByonNodeDomainRepository;
-import javax.transaction.Transactional;
-import org.cloudiator.messages.Byon;
-import org.cloudiator.messages.Byon.ByonNodeRemovedResponse;
-import org.cloudiator.messages.Byon.RemoveByonNodeRequest;
 import org.cloudiator.messages.General.Error;
+import org.cloudiator.messages.Byon.ByonNodeQueryRequest;
+import org.cloudiator.messages.Byon.ByonNodeQueryResponse;
 import org.cloudiator.messaging.MessageInterface;
 import org.cloudiator.messaging.Subscription;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class RemoveByonNodeSubscriber  implements Runnable {
+public class ByonNodeQuerySubscriber implements Runnable {
   private static final Logger LOGGER =
-      LoggerFactory.getLogger(RemoveByonNodeSubscriber.class);
+      LoggerFactory.getLogger(ByonNodeQuerySubscriber.class);
   private final MessageInterface messageInterface;
   private final ByonNodeDomainRepository domainRepository;
   // private final CloudService cloudService;
   private volatile Subscription subscription;
 
   @Inject
-  public RemoveByonNodeSubscriber(MessageInterface messageInterface,
+  public ByonNodeQuerySubscriber(MessageInterface messageInterface,
       ByonNodeDomainRepository domainRepository) {
     this.messageInterface = messageInterface;
     this.domainRepository = domainRepository;
@@ -48,18 +48,15 @@ public class RemoveByonNodeSubscriber  implements Runnable {
 
   @Override
   public void run() {
-    subscription = messageInterface.subscribe(RemoveByonNodeRequest.class,
-        RemoveByonNodeRequest.parser(),
+    subscription = messageInterface.subscribe(ByonNodeQueryRequest.class,
+        ByonNodeQueryRequest.parser(),
         (requestId, request) -> {
           try {
-            String id =  request.getId();
-            LOGGER.debug(String.format("%s retrieved request to delete"
-                    + "byon node with id %s.", this, id));
-            deleteByonNode(id);
-            LOGGER.info("byon node deleted. sending response");
-            messageInterface.reply(requestId,
-                ByonNodeRemovedResponse.newBuilder().build());
-            LOGGER.info("response sent.");
+            String id = request.getId();
+            ByonNodeQueryResponse byonNodeQueryResponse = ByonNodeQueryResponse.newBuilder()
+                .setByonNode(ByonToByonMessageConverter.INSTANCE.apply(domainRepository.findById(id)))
+                .build();
+            messageInterface.reply(requestId, byonNodeQueryResponse);
           } catch (Exception ex) {
             LOGGER.error("Exception occurred.", ex);
             sendErrorResponse(requestId, "Exception occurred: " + ex.getMessage(), Constants.SERVER_ERROR);
@@ -67,13 +64,8 @@ public class RemoveByonNodeSubscriber  implements Runnable {
         });
   }
 
-  @Transactional
-  private void deleteByonNode(String id) {
-    domainRepository.delete(id);
-  }
-
   private final void sendErrorResponse(String messageId, String errorMessage, int errorCode) {
-    messageInterface.reply(ByonNodeRemovedResponse.class, messageId,
+    messageInterface.reply(ByonNodeQueryResponse.class, messageId,
         Error.newBuilder().setCode(errorCode).setMessage(errorMessage).build());
   }
 }
